+++ src/Entity/User.php
@@ -60,6 +60,11 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
         $this->feeds = new ArrayCollection();
     }
 
+    public function __toString(): string
+    {
+        return sprintf('%s (%d)', $this->name ?? '', $this->id ?? -1);
+    }
+
     public function getId(): int
     {
         return $this->id;