+++ src/MessageHandler/FeedNormalizationHandler.php
@@ -2,6 +2,7 @@
 
 namespace App\MessageHandler;
 
+use App\Entity\Event;
 use App\Message\EventMessage;
 use App\Message\FeedNormalizationMessage;
 use App\Service\ContentNormalizer;
@@ -10,13 +11,12 @@ use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 use Symfony\Component\Messenger\MessageBusInterface;
 
 #[AsMessageHandler]
-final class FeedNormalizationHandler
+final readonly class FeedNormalizationHandler
 {
     public function __construct(
-        private readonly ContentNormalizer $contentNormalizer,
-        private readonly MessageBusInterface $messageBus,
-        private readonly TagsNormalizerInterface $tagsNormalizer,
-        private readonly int $excerptMaxLength,
+        private ContentNormalizer $contentNormalizer,
+        private MessageBusInterface $messageBus,
+        private TagsNormalizerInterface $tagsNormalizer,
     ) {
     }
 
@@ -27,15 +27,21 @@ final class FeedNormalizationHandler
         // Tags normalization.
         $item->tags = $this->tagsNormalizer->normalize($item->tags);
 
+        $item->url = $item->url ?? '';
+
         // Url normalization (relative path to full path)
         // @todo should we detect relative paths?
 
         // Content normalizations check up. HTML fixer etc.
         $item->description = $this->contentNormalizer->sanitize($item->description ?? '');
-        if (!empty($item->excerpt)) {
-            $item->excerpt = $this->contentNormalizer->sanitize($item->excerpt);
-            $item->excerpt = $this->contentNormalizer->trimLength($item->excerpt, $this->excerptMaxLength);
+
+        // Set excerpt from description if empty
+        if (empty($item->excerpt) && !empty($item->description)) {
+            $item->excerpt = $item->description;
         }
+        $item->excerpt = $this->contentNormalizer->sanitize($item->excerpt ?? '');
+        $item->excerpt = $this->contentNormalizer->getTextFromHtml($item->excerpt);
+        $item->excerpt = $this->contentNormalizer->trimLength($item->excerpt, Event::EXCERPT_MAX_LENGTH);
 
         $this->messageBus->dispatch(new EventMessage($item));
     }