+++ src/Model/Feed/FeedConfiguration.php
@@ -2,20 +2,69 @@
 
 namespace App\Model\Feed;
 
-final class FeedConfiguration
+final readonly class FeedConfiguration
 {
     public function __construct(
-        public readonly string $type,
-        public readonly string $url,
-        public readonly string $base,
-        public readonly string $timezone,
-        public readonly string $rootPointer,
+        public string $type,
+        public string $url,
+        public string $base,
+        public string $timezone,
+        public string $rootPointer,
         /** @var non-empty-string */
-        public readonly string $dateFormat,
+        public string $dateFormat,
         /** @var array list<string> */
-        public readonly array $mapping = [],
+        public array $mapping = [],
         /** @var array list<string> */
-        public readonly array $defaults = [],
+        public array $defaults = [],
+        public array $clientHeaders = [],
+        public ?FeedPagination $pagination = null,
     ) {
+    }
+
+    public function supportsPagination(): bool
+    {
+        return !(null === $this->pagination) && $this->pagination->supportsPagination();
+    }
+
+    public static function getConfigurationTemplate(): array
+    {
+        return [
+            'type' => 'json',
+            'url' => '',
+            'base' => '',
+            'timezone' => 'Europe/Copenhagen',
+            'rootPointer' => '/-',
+            'dateFormat' => "Y-m-d\TH:i:sP",
+            'pagination' => [
+                'pageParameter' => '',
+                'limitParameter' => '',
+            ],
+            'mapping' => [
+                'id' => 'id',
+                'title' => 'title',
+                'excerpt' => 'excerpt',
+                'description' => 'description',
+                'url' => 'url',
+                'image' => 'image',
+                'occurrences.*.start' => 'occurrences.*.start',
+                'occurrences.*.end' => 'occurrences.*.end',
+                'price' => 'occurrences.*.price',
+            ],
+            'defaults' => [
+                'publicAccess' => true,
+                'location' => [
+                    'name' => '',
+                    'country' => '',
+                    'city' => '',
+                    'postalCode' => null,
+                    'street' => '',
+                    'region' => '',
+                    'url' => '',
+                    'telephone' => '',
+                    'mail' => '',
+                    'disabilityAccess' => true,
+                ],
+            ],
+        ];
     }
 }