+++ src/Model/Feed/FeedItemLocation.php
@@ -16,7 +16,7 @@ final class FeedItemLocation
         public ?string $url = null,
         public ?string $image = null,
         public ?string $region = null,
-        public bool $disabilityAccess = false,
+        public ?bool $disabilityAccess = null,
         public ?FeedItemCoordinates $coordinates = null,
     ) {
     }