+++ CHANGELOG.md
@@ -41,6 +41,7 @@ See [keep a changelog] for information about writing changes to this log.
 - Updated to PHP 8.3 og Symfony 6.4
 - Added ES indexes needed to support endpoints in the API
 - Updated feed import suite to support more feeds
+- Updated feed mapping to support data migration
 
 [keep a changelog]: https://keepachangelog.com/en/1.1.0/
 [unreleased]: https://github.com/itk-dev/event-database-imports/compare/main...develop